+++ src/index/ranker/jelinek_mercer.cpp
@@ -14,20 +14,20 @@ namespace index
 
 const std::string jelinek_mercer::id = "jelinek-mercer";
 
-jelinek_mercer::jelinek_mercer(double lambda) : lambda_{lambda}
+jelinek_mercer::jelinek_mercer(float lambda) : lambda_{lambda}
 {
     /* nothing */
 }
 
-double jelinek_mercer::smoothed_prob(const score_data& sd) const
+float jelinek_mercer::smoothed_prob(const score_data& sd) const
 {
-    double max_likelihood = static_cast<double>(sd.doc_term_count)
+    float max_likelihood = static_cast<float>(sd.doc_term_count)
                             / sd.doc_size;
-    double pc = static_cast<double>(sd.corpus_term_count) / sd.total_terms;
-    return (1.0 - lambda_) * max_likelihood + lambda_ * pc;
+    float pc = static_cast<float>(sd.corpus_term_count) / sd.total_terms;
+    return (1.0f - lambda_) * max_likelihood + lambda_ * pc;
 }
 
-double jelinek_mercer::doc_constant(const score_data&) const
+float jelinek_mercer::doc_constant(const score_data&) const
 {
     return lambda_;
 }
